// src/api/model/productPriceUnit.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Backend API
 * Internal API supplying product/pricing data (bricks and construction products), transport costs and surcharges, plus dashboard configuration. Session-cookie auth; mutations and account management require authentication.
 * OpenAPI spec version: 1.0.0
 */

export type ProductPriceUnit = typeof ProductPriceUnit[keyof typeof ProductPriceUnit];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductPriceUnit = {
  m2: 'm2',
  бр: 'бр.',
} as const;